+++ example/lib/bloc/theme_bloc.dart
@@ -21,3 +21,7 @@ class ChangeThemeColor extends ThemeEvent {
   ChangeThemeColor(this.color);
 }
 
+
+
+
+