+++ example/lib/widgets/themed_svg_icon.dart
@@ -67,3 +67,7 @@ class ThemedSvgIcon extends StatelessWidget {
   }
 }
 
+
+
+
+